requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
langchain==0.3.27
langchain-groq==0.3.6
pydantic==2.11.7
python-multipart==0.0.6
cors==1.0.1 